requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-generativeai
pylint
